feat(dashboard): vista resumen visual + mejoras KPI y semáforo
- Vista toggle Resumen / Tabla
- SemaforoBar: barra global de estado del portafolio por pedidos únicos
- GroupCards: cards por asesor (gerente) o por cliente (asesor) con mini barra
- ProximasEntregas: próximos 14 días ordenados por urgencia
- KPIs: vencidos/urgentes cuentan pedidos únicos (no líneas)
- getSemaforoLevel helper con niveles: vencido, hoy, urgente, pronto, ok, sin_fecha
- Paginación sin claves duplicadas
- schema.sql: cliente_sucursal bigint

+++ supabase/schema.sql
@@ -88,7 +88,7 @@ create table public.pedidos (
 
   -- Cliente (de informe_pendientes col C-E)
   cliente_nit           bigint not null,
-  cliente_sucursal      integer not null default 0,
+  cliente_sucursal      bigint not null default 0,
   cliente_nombre        text not null,
 
   -- Producto (de informe_pendientes col F-H)
